move parameterbase and cache to own modules
