+++ qcodes/tests/helpers/test_on_off_mapping.py
@@ -1,6 +1,6 @@
 import pytest
 
-from qcodes.parameters.parameter import invert_val_mapping
+from qcodes.parameters import invert_val_mapping
 from qcodes.utils.helpers import create_on_off_val_mapping
